requirements.txt
aiosqlite==0.21.0
azure-ai-projects==2.0.0b2
azure-identity==1.25.1
fastapi==0.124.0
langchain==1.1.2
langchain-core==1.1.1
langchain-openai==1.1.0
langgraph==1.0.4
pydantic==2.12.5
python-dotenv==1.2.1
python-multipart==0.0.20
uvicorn==0.38.0
